backend/pyproject.toml
[tool.poetry]
name = "money-trader-backend"
version = "0.1.0"
description = "Money Trader FastAPI backend"
authors = ["Money Trader Team <team@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pydantic = "^1.10.14"
python-dotenv = "^1.0.1"
celery = "^5.3.4"
redis = "^5.0.3"
httpx = "^0.27.0"
argon2-cffi = "^23.1.0"
email-validator = "^2.1.0.post1"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
